Refactor admin UI into modular dark theme
