+++ server/admin/__init__.py
@@ -1,0 +1,3 @@
+from .router import create_admin_router
+
+__all__ = ['create_admin_router']
