+++ server/admin/views/__init__.py
@@ -1,0 +1,2 @@
+# View registration modules for the admin panel
+__all__ = []
