+++ templates/admin/base.html
@@ -1,0 +1,246 @@
+<!doctype html>
+<html lang="ru">
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <title>{% block title %}Админка MentorMatch{% endblock %}</title>
+  <style>
+    :root {
+      color-scheme: dark;
+    }
+    * {
+      box-sizing: border-box;
+    }
+    body {
+      margin: 0;
+      font-family: 'Inter', system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+      background: #020617;
+      color: #e2e8f0;
+      min-height: 100vh;
+      display: flex;
+      flex-direction: column;
+    }
+    a {
+      color: #38bdf8;
+      text-decoration: none;
+    }
+    a:hover {
+      color: #f8fafc;
+    }
+    .layout {
+      flex: 1 1 auto;
+      max-width: 1200px;
+      margin: 0 auto;
+      padding: 32px 24px 48px;
+      display: flex;
+      flex-direction: column;
+      gap: 32px;
+    }
+    header {
+      display: flex;
+      flex-wrap: wrap;
+      justify-content: space-between;
+      align-items: center;
+      gap: 16px;
+    }
+    header h1 {
+      font-size: 1.75rem;
+      font-weight: 600;
+      margin: 0;
+      letter-spacing: -0.02em;
+    }
+    .surface {
+      background: rgba(15, 23, 42, 0.72);
+      border: 1px solid rgba(148, 163, 184, 0.15);
+      border-radius: 18px;
+      padding: 24px;
+      backdrop-filter: blur(18px);
+      box-shadow: 0 18px 60px rgba(2, 6, 23, 0.45);
+    }
+    .surface + .surface {
+      margin-top: 0;
+    }
+    .muted {
+      color: rgba(148, 163, 184, 0.85);
+      font-size: 0.95rem;
+    }
+    .btn {
+      appearance: none;
+      border: none;
+      outline: none;
+      border-radius: 999px;
+      padding: 10px 18px;
+      font-weight: 600;
+      font-size: 0.95rem;
+      cursor: pointer;
+      transition: transform 120ms ease, box-shadow 120ms ease, background 120ms ease;
+      background: linear-gradient(135deg, #22d3ee, #2563eb);
+      color: #0f172a;
+      box-shadow: 0 12px 30px rgba(37, 99, 235, 0.35);
+    }
+    .btn:hover {
+      transform: translateY(-1px);
+      box-shadow: 0 18px 45px rgba(37, 99, 235, 0.45);
+    }
+    .btn-secondary {
+      background: rgba(51, 65, 85, 0.65);
+      color: #e2e8f0;
+      box-shadow: none;
+    }
+    .btn-secondary:hover {
+      background: rgba(71, 85, 105, 0.8);
+    }
+    .btn-danger {
+      background: linear-gradient(135deg, #f87171, #ef4444);
+      color: #0f172a;
+      box-shadow: 0 12px 30px rgba(248, 113, 113, 0.3);
+    }
+    .badge {
+      padding: 4px 10px;
+      border-radius: 999px;
+      background: rgba(148, 163, 184, 0.12);
+      color: rgba(226, 232, 240, 0.9);
+      font-size: 0.75rem;
+      font-weight: 600;
+      text-transform: uppercase;
+      letter-spacing: 0.08em;
+    }
+    input[type="text"],
+    input[type="number"],
+    input[type="email"],
+    textarea,
+    select {
+      width: 100%;
+      border-radius: 12px;
+      border: 1px solid rgba(148, 163, 184, 0.25);
+      background: rgba(15, 23, 42, 0.7);
+      color: #e2e8f0;
+      padding: 12px 14px;
+      font-size: 0.95rem;
+      transition: border 120ms ease, box-shadow 120ms ease;
+    }
+    input:focus,
+    textarea:focus,
+    select:focus {
+      border-color: rgba(56, 189, 248, 0.8);
+      box-shadow: 0 0 0 3px rgba(56, 189, 248, 0.25);
+      outline: none;
+    }
+    textarea {
+      resize: vertical;
+      min-height: 120px;
+    }
+    label span.title {
+      display: block;
+      font-weight: 600;
+      margin-bottom: 6px;
+    }
+    table {
+      width: 100%;
+      border-collapse: collapse;
+    }
+    table thead tr {
+      text-transform: uppercase;
+      font-size: 0.75rem;
+      letter-spacing: 0.08em;
+      color: rgba(148, 163, 184, 0.85);
+    }
+    table tbody tr {
+      border-top: 1px solid rgba(148, 163, 184, 0.12);
+    }
+    table td, table th {
+      padding: 14px 12px;
+      text-align: left;
+      vertical-align: top;
+    }
+    .stack {
+      display: flex;
+      flex-direction: column;
+      gap: 16px;
+    }
+    .grid {
+      display: grid;
+      gap: 20px;
+    }
+    .grid.two {
+      grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
+    }
+    form.actions {
+      display: flex;
+      gap: 12px;
+      flex-wrap: wrap;
+      align-items: center;
+    }
+    nav.tabs {
+      display: flex;
+      gap: 12px;
+      flex-wrap: wrap;
+    }
+    nav.tabs a {
+      padding: 10px 18px;
+      border-radius: 999px;
+      background: rgba(30, 41, 59, 0.6);
+      color: rgba(226, 232, 240, 0.85);
+      font-weight: 600;
+      transition: background 120ms ease, color 120ms ease;
+    }
+    nav.tabs a.active {
+      background: rgba(56, 189, 248, 0.15);
+      color: #e0f2fe;
+    }
+    nav.tabs a:hover {
+      background: rgba(56, 189, 248, 0.25);
+      color: #f8fafc;
+    }
+    .message {
+      padding: 14px 18px;
+      border-radius: 14px;
+      border: 1px solid rgba(56, 189, 248, 0.25);
+      background: rgba(56, 189, 248, 0.1);
+      color: #e0f2fe;
+    }
+    .chip-list {
+      display: flex;
+      gap: 8px;
+      flex-wrap: wrap;
+    }
+    .chip {
+      padding: 6px 12px;
+      border-radius: 999px;
+      background: rgba(148, 163, 184, 0.14);
+      font-size: 0.85rem;
+    }
+    .form-grid {
+      display: grid;
+      gap: 20px;
+    }
+    .form-grid.two {
+      grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
+    }
+    .section-title {
+      font-size: 1.1rem;
+      font-weight: 600;
+      margin-bottom: 12px;
+    }
+    @media (max-width: 720px) {
+      .layout {
+        padding: 24px 18px 36px;
+      }
+    }
+  </style>
+  {% block extra_head %}{% endblock %}
+</head>
+<body>
+  <div class="layout">
+    <header>
+      <div>
+        <h1>MentorMatch · админ-панель</h1>
+        {% block header_meta %}{% endblock %}
+      </div>
+      {% block header_actions %}{% endblock %}
+    </header>
+
+    {% block content %}{% endblock %}
+  </div>
+</body>
+</html>
