+++ templates/admin/view_topic.html
@@ -1,0 +1,83 @@
+{% extends 'admin/base.html' %}
+{% block title %}Тема #{{ topic.id }} · {{ topic.title }}{% endblock %}
+{% block header_meta %}
+<p class="muted">Подробности темы и рекомендации</p>
+{% endblock %}
+{% block content %}
+  <section class="surface stack">
+    {% if msg %}
+      <div class="message">{{ msg }}</div>
+    {% endif %}
+    <div class="stack" style="gap:18px;">
+      <div>
+        <div class="badge">Тема #{{ topic.id }}</div>
+        <h2 style="margin:8px 0 0; font-size:1.5rem;">{{ topic.title or 'Без названия' }}</h2>
+      </div>
+      <div class="muted">Автор: {{ topic.author }}</div>
+      <div class="muted">Роль: {{ topic.seeking_role }}{% if topic.direction %}, направление {{ topic.direction }}{% endif %}</div>
+      <div class="stack">
+        <h3 class="section-title">Описание</h3>
+        <p>{{ topic.description or '—' }}</p>
+      </div>
+      <div class="stack">
+        <h3 class="section-title">Ожидаемые результаты</h3>
+        <p>{{ topic.expected_outcomes or '—' }}</p>
+      </div>
+      <div class="stack">
+        <h3 class="section-title">Требуемые навыки</h3>
+        <p>{{ topic.required_skills or '—' }}</p>
+      </div>
+      <div class="topic-actions">
+        <a class="btn-secondary" href="/edit-topic/{{ topic.id }}">Редактировать</a>
+        <a class="btn-secondary" href="/add-role?topic_id={{ topic.id }}">Добавить роль</a>
+      </div>
+    </div>
+  </section>
+
+  <section class="surface stack">
+    <h3 class="section-title">Роли темы</h3>
+    <div class="role-grid">
+      {% for role in roles %}
+        <div class="surface" style="padding:18px;">
+          <div class="stack" style="gap:12px;">
+            <div style="display:flex;justify-content:space-between;align-items:center;">
+              <strong>{{ role.name or ('Роль #' ~ role.id) }}</strong>
+              <a class="muted" href="/role/{{ role.id }}">Подробнее →</a>
+            </div>
+            <div class="muted">Навыки: {{ role.required_skills or '—' }}</div>
+            <div class="muted">Вместимость: {{ role.capacity or '—' }}</div>
+            {% if role.approved_student_user_id %}
+              <div class="muted">Назначен студент ID {{ role.approved_student_user_id }}</div>
+            {% endif %}
+          </div>
+        </div>
+      {% else %}
+        <p class="muted">Роли ещё не созданы.</p>
+      {% endfor %}
+    </div>
+  </section>
+
+  <section class="surface stack">
+    <h3 class="section-title">Подходящие руководители</h3>
+    <div class="stack">
+      {% for cand in supervisor_candidates %}
+        <div class="surface" style="padding:18px;">
+          <div class="stack" style="gap:10px;">
+            <div style="display:flex;justify-content:space-between;align-items:center;">
+              <strong>{{ cand.full_name }}</strong>
+              <span class="badge">Ранг {{ cand.rank or '—' }}</span>
+            </div>
+            <div class="muted">{{ cand.position or '—' }} · {{ cand.degree or '—' }}</div>
+            <div class="muted">Интересы: {{ cand.interests or '—' }}</div>
+            <details>
+              <summary class="muted" style="cursor:pointer;">Шаблон сообщения</summary>
+              <p style="margin-top:8px;">{{ cand.default_message }}</p>
+            </details>
+          </div>
+        </div>
+      {% else %}
+        <p class="muted">Рекомендаций пока нет.</p>
+      {% endfor %}
+    </div>
+  </section>
+{% endblock %}
