--- templates/edit_supervisor.html
@@ -1,41 +1,0 @@
-<!doctype html>
-<html lang='ru'>
-<head>
-  <meta charset='utf-8' />
-  <meta name='viewport' content='width=device-width, initial-scale=1' />
-  <title>Изменить руководителя</title>
-  <style>
-    body { font-family: system-ui,-apple-system,Segoe UI,Roboto,Arial,sans-serif; margin: 20px; line-height: 1.4; }
-    .card { border: 1px solid #e2e8f0; border-radius: 12px; padding: 16px; max-width: 820px; }
-    label { display:block; font-size:0.9rem; margin-bottom:8px; }
-    input[type='text'], input[type='email'], input[type='number'], textarea, select { width: 100%; padding: 8px 10px; border: 1px solid #cbd5e1; border-radius: 8px; font-size: 14px; }
-    textarea { min-height: 80px; }
-    button { background: #2563eb; color: white; border: 0; padding: 10px 14px; border-radius: 8px; cursor: pointer; }
-    button:hover { background: #1e40af; }
-    .row { display:flex; gap: 12px; align-items:center; }
-    .muted { color: #64748b; font-size: 0.9rem; }
-    a { color:#2563eb; }
-  </style>
-</head>
-<body>
-  <h2>Изменить руководителя #{{ sup.id }}</h2>
-  <div class='card'>
-    <form method='post' action='/update-supervisor'>
-      <input type='hidden' name='user_id' value='{{ sup.id }}'>
-      <label>ФИО <input type='text' name='full_name' value='{{ sup.full_name }}' required></label>
-      <label>Email <input type='text' name='email' value='{{ sup.email or "" }}'></label>
-      <label>Username (tg) <input type='text' name='username' value='{{ sup.username or "" }}'></label>
-      <label>Должность <input type='text' name='position' value='{{ sup.position or "" }}'></label>
-      <label>Учёная степень <input type='text' name='degree' value='{{ sup.degree or "" }}'></label>
-      <label>Лимит студентов <input type='number' name='capacity' value='{{ sup.capacity or "" }}' min='0'></label>
-      <label>Интересы <textarea name='interests'>{{ sup.interests or '' }}</textarea></label>
-      <label>Требования <textarea name='requirements'>{{ sup.requirements or '' }}</textarea></label>
-      <div class='row'>
-        <button type='submit'>Сохранить</button>
-        <a href='/?kind=supervisors' class='muted'>Отмена</a>
-      </div>
-    </form>
-  </div>
-</body>
-</html>
-
