--- templates/edit_user.html
@@ -1,44 +1,0 @@
-<!doctype html>
-<html lang='ru'>
-<head>
-  <meta charset='utf-8' />
-  <meta name='viewport' content='width=device-width, initial-scale=1' />
-  <title>Изменить пользователя</title>
-  <style>
-    body { font-family: system-ui,-apple-system,Segoe UI,Roboto,Arial,sans-serif; margin: 20px; line-height: 1.4; }
-    .card { border: 1px solid #e2e8f0; border-radius: 12px; padding: 16px; max-width: 720px; }
-    label { display:block; font-size:0.9rem; margin-bottom:8px; }
-    input[type='text'], input[type='email'], input[type='number'], textarea, select { width: 100%; padding: 8px 10px; border: 1px solid #cbd5e1; border-radius: 8px; font-size: 14px; }
-    button { background: #2563eb; color: white; border: 0; padding: 10px 14px; border-radius: 8px; cursor: pointer; }
-    button:hover { background: #1e40af; }
-    .row { display:flex; gap: 12px; align-items:center; }
-    .muted { color: #64748b; font-size: 0.9rem; }
-    a { color:#2563eb; }
-  </style>
-  </head>
-  <body>
-    <h2>Изменить пользователя #{{ user.id }}</h2>
-    <div class='card'>
-      <form method='post' action='/update-user'>
-        <input type='hidden' name='user_id' value='{{ user.id }}'>
-        <label>ФИО <input type='text' name='full_name' value='{{ user.full_name }}' required></label>
-        <label>Email <input type='text' name='email' value='{{ user.email or "" }}'></label>
-        <label>Username (tg) <input type='text' name='username' value='{{ user.username or "" }}'></label>
-        <label>Роль
-          <select name='role'>
-            <option value='student' {{ 'selected' if user.role=='student' else '' }}>student</option>
-            <option value='supervisor' {{ 'selected' if user.role=='supervisor' else '' }}>supervisor</option>
-            <option value='admin' {{ 'selected' if user.role=='admin' else '' }}>admin</option>
-          </select>
-        </label>
-        <label><input type='checkbox' name='consent_personal' {{ 'checked' if user.consent_personal else '' }}> Согласие на обработку ПДн</label>
-        <label><input type='checkbox' name='consent_private' {{ 'checked' if user.consent_private else '' }}> Согласие на публикации в приватных чатах</label>
-        <div class='row'>
-          <button type='submit'>Сохранить</button>
-          <a href='/' class='muted'>Отмена</a>
-        </div>
-      </form>
-    </div>
-  </body>
-</html>
-
